+++ modules/fetchdata.js
@@ -1,0 +1,14 @@
+import commentPopup from './commentPopup.js';
+import displayCard from './displayCard.js';
+
+const API = 'https://api.tvmaze.com/shows';
+
+export default async () => {
+  await fetch(API)
+    .then((response) => response.json())
+    .then((json) => json)
+    .then((data) => {
+      displayCard(data);
+      commentPopup(data);
+    });
+};
