+++ src/index.html
@@ -34,6 +34,7 @@
     <main class="main-section">
       <!-- cards willbe render dynamically using fetch with Api -->
     </main>
+
     <div id="comment-section" class="comment-show">
       <section id="comment-popup" class="popup">
         <div id="comment-info" class="movie-info"></div>
