+++ src/style.css
@@ -140,6 +140,7 @@ h3 {
   background-color: rgb(215, 255, 242);
 }
 
+/*  comment popup */
 form > input,
 textarea {
   outline: none;
